+++ src/test/kotlin/net/pwall/json/schema/parser/ParserTest.kt
@@ -2,7 +2,7 @@
  * @(#) ParserTest.kt
  *
  * json-kotlin-schema Kotlin implementation of JSON Schema
- * Copyright (c) 2020 Peter Wall
+ * Copyright (c) 2020, 2021 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal